+++ Packages/webxr/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Native XRSubsystem for WebXR Headset/Device tracking.
 - WebXRCameraSettings for updating Camera settings when switching XR modes.
 
+### Removed
+- Support for Unity versions older than 2020.3.6f1.
+
 ## [0.19.0] - 2023-11-22
 ### Added
 - WebXRHandData pointerPosition and pointerRotation (WebXR targetRaySpace for hands).